Move common function to utils
The function which checks whether the operator should delete a
stale route is common to all three operators and should move to
the utils module
