+++ controllers/runtimecomponent_controller.go
@@ -432,7 +432,7 @@ func (r *RuntimeComponentReconciler) Reconcile(ctx context.Context, req ctrl.Req
 		r.ManageError(err, common.StatusConditionTypeReconciled, instance)
 	} else if ok {
 		if instance.Spec.Expose != nil && *instance.Spec.Expose {
-			if shouldDeleteRoute(ba) {
+			if appstacksutils.ShouldDeleteRoute(ba) {
 				reqLogger.Info("Custom hostname has been removed from route, deleting and recreating the route")
 				route := &routev1.Route{ObjectMeta: defaultMeta}
 				err = r.DeleteResource(route)
@@ -640,20 +640,4 @@ func (r *RuntimeComponentReconciler) SetupWithManager(mgr ctrl.Manager) error {
 
 func getMonitoringEnabledLabelName(ba common.BaseComponent) string {
 	return "monitor." + ba.GetGroupName() + "/enabled"
-}
-
-// If a custome hostname was previously set, but is now not set, any previous
-// route needs to be deleted, as the host in a route cannot be unset
-// and the default generated hostname is difficult to manually recreate
-func shouldDeleteRoute(ba common.BaseComponent) bool {
-	rh := ba.GetStatus().GetReferences()[common.StatusReferenceRouteHost]
-	if rh != "" {
-		// The host was previously set.
-		// If the host is now empty, delete the old route
-		rt := ba.GetRoute()
-		if rt == nil || (rt.GetHost() == "" && common.Config[common.OpConfigDefaultHostname] == "") {
-			return true
-		}
-	}
-	return false
 }
