+++ utils/utils.go
@@ -1853,3 +1853,19 @@ func GetServiceAccountName(ba common.BaseComponent) string {
 
 	return name
 }
+
+// If a custome hostname was previously set, but is now not set, any previous
+// route needs to be deleted, as the host in a route cannot be unset
+// and the default generated hostname is difficult to manually recreate
+func ShouldDeleteRoute(ba common.BaseComponent) bool {
+	rh := ba.GetStatus().GetReferences()[common.StatusReferenceRouteHost]
+	if rh != "" {
+		// The host was previously set.
+		// If the host is now empty, delete the old route
+		rt := ba.GetRoute()
+		if rt == nil || (rt.GetHost() == "" && common.Config[common.OpConfigDefaultHostname] == "") {
+			return true
+		}
+	}
+	return false
+}
